When using Rails to truncate strings, you may end up with strings that are still too long for their container or are not as long as they could be. You can get a prettier result using stylesheets.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it
When using Rails to truncate strings, you may end up with strings that are still too long for their container or are not as long as they could be. You can get a prettier result using stylesheets.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it.This is a new commit message that I am telling to my shippable app to take only a few characters and do the build. What basically i have done here is added a new commit message that is really long and experimenting on it